tes: restructure code
