+++ model/checkin.go
@@ -1,0 +1,7 @@
+package model
+
+type Checkin struct {
+	Identifier string
+	Type       string
+	LocationID int
+}
