+++ repository/db.go
@@ -10,4 +10,5 @@ import (
 type DB interface {
 	io.Closer
 	ListLocation(ctx context.Context) ([]*model.Location, error)
+	InputCheckin(ctx context.Context) ([]*model.CheckInOutRequest, error)
 }
